Alerts: Support cheers and some channel point redemptions
For now the Alerts triggered are very hardcoded. This will change at
some point in the future.

EventSub backend now reacts properly to both bits being cheered on a
channel and channel points reward being redeemed. Event is passed to the
Alerts widget, which handles it further.

This change is mostly done as a POC to recognize what should be done to
properly support these behaviors, either at LukeBot or at some further
project. Many things (especially how the alert widget behaves) are
subject to change.