Move user switching to CLIMessageProxy
Getting/Setting current user is now a part of CLIMessageProxy. This
allows to implement it per-connection on ServerCLI side.

To help with thread safety, CLIMessageProxy references in CLIProcessors
were moved to be passed as method argument in handlers.

Also used this opportunity to rename "user select" to "user switch"
command. Makes more sense in what this function performs.